[Refactor] 코드 컨포넌트 파일 분리 & 주석 수정
